Add support for Ruby syntax highlighting
