+++ .prettierrc.json
@@ -1,0 +1,13 @@
+{
+  "printWidth": 80,
+  "arrowParens": "avoid",
+  "trailingComma": "es5",
+  "overrides": [
+    {
+      "files": "*.{jsx,tsx}",
+      "options": {
+        "printWidth": 100
+      }
+    }
+  ]
+}
