+++ package.json
@@ -44,6 +44,7 @@
     "react-helmet": "^5.2.1",
     "react-instantsearch-dom": "^5.7.0",
     "react-live": "^2.1.2",
+    "refractor": "^2.10.0",
     "styled-components": "^4.3.2",
     "styled-icons": "^8.3.0",
     "system-components": "^3.0.3"
